requirements.txt
pandas
openai
vllm
torch
tqdm
pyyaml
scikit-learn
numpy
ollama
